[#1268] move http services to http_tracker_core package
