+++ src/servers/http/v1/mod.rs
@@ -5,4 +5,3 @@
 pub mod extractors;
 pub mod handlers;
 pub mod routes;
-pub mod services;
